+++ akshare/stock/stock_zh_a_sina.py
@@ -1,20 +1,17 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2025/3/20 19:00
+Date: 2026/1/9 22:00
 Desc: 新浪财经-A股-实时行情数据和历史行情数据(包含前复权和后复权因子)
 https://finance.sina.com.cn/realstock/company/sh689009/nc.shtml
 """
 
 import json
 import re
-from io import StringIO
 
 import pandas as pd
 import py_mini_racer
 import requests
-from bs4 import BeautifulSoup
-
 
 from akshare.stock.cons import (
     zh_sina_a_stock_payload,
@@ -24,7 +21,7 @@ from akshare.stock.cons import (
     hk_js_decode,
     zh_sina_a_stock_hfq_url,
     zh_sina_a_stock_qfq_url,
-    zh_sina_a_stock_amount_page_url,
+    zh_sina_a_stock_amount_url
 )
 from akshare.utils import demjson
 from akshare.utils.tqdm import get_tqdm
@@ -196,32 +193,12 @@ def stock_zh_a_daily(
     except:  # noqa: E722
         pass
     data_df = data_df.astype("float")
-
-    amount_data_dfs = []
-    r = requests.get(zh_sina_a_stock_amount_page_url.format(symbol[2:]))
-    soup = BeautifulSoup(r.text, "html.parser")
-    th = soup.find(name="th", string="持有股数")
-    # 如果完全没有数据，可能找不到对应的<th>。
-    if th is not None:
-        for table in th.find_parent("table").find_parent("table").find_all("table"):
-            # 这里的<table>也有可能没有数据，需要跳过。
-            if table.find(name="th", string="持有股数") is None:
-                continue
-            amount_data_dfs.append(
-                pd.read_html(StringIO(str(table)))[0]
-                .set_axis(["date", "outstanding_share"], axis="columns")
-            )
-    if len(amount_data_dfs) == 0:
-        amount_data_df = pd.DataFrame(columns=["outstanding_share"], dtype=float)
-    else:
-        amount_data_df = pd.concat(amount_data_dfs, axis="index", ignore_index=True)
-        amount_data_df.index = pd.to_datetime(amount_data_df.date)
-        amount_data_df.sort_index(inplace=True)
-        del amount_data_df["date"]
-        amount_data_df["outstanding_share"] = (
-            amount_data_df["outstanding_share"].str[:-2].astype(float)
-        )
-
+    r = requests.get(zh_sina_a_stock_amount_url.format(symbol, symbol))
+    amount_data_json = demjson.decode(r.text[r.text.find("["): r.text.rfind("]") + 1])
+    amount_data_df = pd.DataFrame(amount_data_json)
+    amount_data_df.columns = ["date", "outstanding_share"]
+    amount_data_df.index = pd.to_datetime(amount_data_df.date)
+    del amount_data_df["date"]
     temp_df = pd.merge(
         data_df, amount_data_df, left_index=True, right_index=True, how="outer"
     )
@@ -452,7 +429,7 @@ def stock_zh_a_minute(
         need_df.index = pd.to_datetime(need_df["date"])
         stock_zh_a_daily_qfq_df = stock_zh_a_daily(symbol=symbol, adjust="qfq")
         stock_zh_a_daily_qfq_df.index = pd.to_datetime(stock_zh_a_daily_qfq_df["date"])
-        result_df = stock_zh_a_daily_qfq_df.iloc[-len(need_df) :, :]["close"].astype(
+        result_df = stock_zh_a_daily_qfq_df.iloc[-len(need_df):, :]["close"].astype(
             float
         ) / need_df["close"].astype(float)
         temp_df.index = pd.to_datetime(temp_df["date"])
@@ -477,7 +454,7 @@ def stock_zh_a_minute(
         need_df.index = pd.to_datetime(need_df["date"])
         stock_zh_a_daily_hfq_df = stock_zh_a_daily(symbol=symbol, adjust="hfq")
         stock_zh_a_daily_hfq_df.index = pd.to_datetime(stock_zh_a_daily_hfq_df["date"])
-        result_df = stock_zh_a_daily_hfq_df.iloc[-len(need_df) :, :]["close"].astype(
+        result_df = stock_zh_a_daily_hfq_df.iloc[-len(need_df):, :]["close"].astype(
             float
         ) / need_df["close"].astype(float)
         temp_df.index = pd.to_datetime(temp_df["date"])
@@ -495,9 +472,9 @@ def stock_zh_a_minute(
 
 if __name__ == "__main__":
     stock_zh_a_daily_hfq_df_one = stock_zh_a_daily(
-        symbol="sz000001",
+        symbol="sz000002",
         start_date="19910403",
-        end_date="20231027",
+        end_date="20260109",
         adjust="hfq",
     )
     print(stock_zh_a_daily_hfq_df_one)