+++ docs/changelog.md
@@ -89,6 +89,11 @@
 
 ## 更新说明详情
 
+1.18.9 fix: fix fx_quote_baidu interface
+
+	1. 修复 fx_quote_baidu 接口
+	2. 修复 stock_zh_a_daily 接口
+
 1.18.8 fix: fix stock_zh_valuation_comparison_em interface
 
 	1. 修复 stock_zh_valuation_comparison_em 接口
@@ -6056,6 +6061,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.18.9 fix: fix fx_quote_baidu interface
 
 1.18.8 fix: fix stock_zh_valuation_comparison_em interface